Migrate SchemaExplorer
